requirements.txt
tabulate
psutil
altair==2.3.0
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrdict==2.0.0
attrs==19.1.0
backcall==0.1.0
boto3==1.9.188
botocore==1.12.188
bravado==10.4.1
bravado-core==5.13.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
docutils==0.14
-e git+https://github.com/facebookresearch/EGG.git#egg=EGG
entrypoints==0.3
future==0.17.1
gitdb2==2.0.5
GitPython==2.1.11
hyperopt==0.1.1
idna==2.8
importlib-metadata==0.18
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonref==0.2
jsonschema==3.0.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
monotonic==1.5
more-itertools==7.1.0
msgpack-python==0.5.6
neptune-client==0.3.0
neptune-contrib==0.9.1
networkx==2.3
numpy==1.16.4
nvidia-ml-py3==7.352.0
oauthlib==3.0.2
packaging==19.0
pandas==0.24.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.12.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pymongo==3.8.0
pyparsing==2.4.0
pyrsistent==0.15.3
pytest==5.0.1
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
retrying==1.3.3
s3transfer==0.2.1
scikit-learn==0.21.2
scikit-optimize==0.5.2
scikit-plot==0.3.7
scipy==1.3.0
seaborn==0.9.0
simplejson==3.16.0
six==1.12.0
smmap2==2.0.5
swagger-spec-validator==2.4.3
toolz==0.10.0
torch==1.1.0
torchvision==0.3.0
traitlets==4.3.2
typing-extensions==3.7.4
urllib3==1.25.3
vega==2.4.0
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.5.2
